// libs/shared/src/utils/enums/errors.enum.ts
export enum HttpError {
  GENERIC_ERROR = 'GENERIC_ERROR',
  INTERNAL_SERVER_ERROR = 'Internal Server Error',
  BAD_REQUEST = 'Bad Request',
  UNAUTHORIZED = 'Unauthorized',
  FORBIDDEN = 'Forbidden',
  NOT_FOUND = 'Not Found',
  FOUND = 'Found',
  VALIDATION_ERROR = 'Validation Error',
  DATABASE_ERROR = 'Database Error',
  AUTHENTICATION_ERROR = 'Authentication Error',
  AUTHORIZATION_ERROR = 'Authorization Error',
}